third_party/WebKit/LayoutTests/http/tests/media/video-load-metadata-decode-error.html

<!DOCTYPE html>
<title>This test case simulates a decode error after loading meta data of a video.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../../media-resources/media-file.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    var watcher = new EventWatcher(t, video, ["loadedmetadata", "error"]);
    watcher.wait_for(["loadedmetadata", "error"]).then(t.step_func_done(function() {
        assert_equals(video.networkState, HTMLMediaElement.NETWORK_IDLE);
    }));
    var movie = "resources/test.ogv";
    var type = mimeTypeForExtension(movie.split(".").pop());
    video.src = "video-load-metadata-decode-error.cgi?name=" + movie + "&type=" + type;
    video.play();
});
</script>
